Squashed 'resources/hpxml-measures/' changes from 7bf7086324f..d7270930d52
d7270930d52 Merge pull request #1385 from jmarrec/1384_WWR_Fallback_add_windows
38094d23ac8 Apply suggestions from code review
9528831dd85 Merge pull request #1388 from NREL/fix-geometry-fin-attic
28d2a7ee867 Minor docs clarification.
5f4df82638c Merge pull request #1389 from NREL/fix-neighbor-building-azimuth-error
b47fa99d8e6 Latest results.
80507d7c1a5 Generalize test logic a bit.
9b7e1b07da3 Typo [ci skip]
e47c2f0767a Proposed fix.
9ffb4b65410 Add test file demonstrating the error.
f904dffb95d Ability to debug geometry in build measure tests.
9782db891aa Same attic over garage rules for conditioned attic.
2b668ce1ccd Fix RTD?
b7c37082768 #1384 - Fallback to a WWR calculation when window placement fails

git-subtree-dir: resources/hpxml-measures
git-subtree-split: d7270930d523623e375276187f5504d10829f07d

+++ BuildResidentialHPXML/tests/build_residential_hpxml_test.rb
@@ -10,10 +10,11 @@ require 'fileutils'
 class BuildResidentialHPXMLTest < MiniTest::Test
   def setup
     @output_path = File.join(File.dirname(__FILE__), 'extra_files')
+    @model_save = false # true helpful for debugging, i.e., can render osm in 3D
   end
 
   def teardown
-    FileUtils.rm_rf(@output_path)
+    FileUtils.rm_rf(@output_path) if !@model_save
   end
 
   def test_workflows
@@ -288,6 +289,7 @@ class BuildResidentialHPXMLTest < MiniTest::Test
 
         # Apply measure
         success = apply_measures(measures_dir, measures, runner, model)
+        model.save(File.absolute_path(File.join(@output_path, hpxml_file.gsub('.xml', '.osm')))) if @model_save
 
         _test_measure(runner, expected_errors[hpxml_file], expected_warnings[hpxml_file])
